move utility functions to util.hpp
