+++ test/unit/math/rev/prob/multi_normal_prec2_test.cpp
@@ -5,6 +5,7 @@
 #include <test/unit/math/rev/prob/expect_eq_diffs.hpp>
 #include <test/unit/math/prim/prob/agrad_distributions_multi_normal_multi_row.hpp>
 #include <test/unit/math/prim/prob/agrad_distributions_multi_normal.hpp>
+#include <test/unit/math/util.hpp>
 #include <vector>
 #include <string>
