+++ test/unit/math/rev/prob/test_gradients_multi_normal.hpp
@@ -4,26 +4,7 @@
 #include <vector>
 #include <iomanip>
 #include <stdexcept>
-
-std::vector<stan::math::var> get_vvar(std::vector<double> vd) {
-  size_t vd_size = vd.size();
-  std::vector<stan::math::var> vv;
-  vv.reserve(vd_size);
-  for (size_t i = 0; i < vd_size; i++)
-    vv.push_back(vd[i]);
-  return vv;
-}
-
-std::vector<double> vdouble_from_vvar(std::vector<stan::math::var> vv) {
-  size_t vv_size = vv.size();
-  std::vector<double> vd;
-  vd.reserve(vv_size);
-  for (size_t i = 0; i < vv_size; i++)
-    vd.push_back(vv[i].val());
-  return vd;
-}
-
-std::vector<double> vdouble_from_vvar(std::vector<double> vv) { return vv; }
+#include <test/unit/math/util.hpp>
 
 template <typename F, typename T_y, typename T_mu, typename T_sigma>
 std::vector<double> finite_diffs_multi_normal(
